requirements.txt
aiogram==3.*
fastapi==0.103.*
uvicorn==0.23.*
httpx==0.25.*
jinja2==3.*
python-dotenv==1.*
pyyaml
playwright
deep-translator